+++ app/Http/Livewire/Pertanian.php
@@ -5,6 +5,7 @@ use App\Models\IdentitasSekolah;
 use App\Models\Blog;
 use Te7aHoudini\LaravelTrix\Models\TrixAttachment;
 use Livewire\Component;
+use App\Models\ImgCarIdenSekolah;
 
 class Pertanian extends Component
 {
@@ -67,9 +68,12 @@ class Pertanian extends Component
     {
         // dd(request('blog-trixFields'));
         $trad = TrixAttachment::where('is_pending', 1)->get();
-        $this->t_content_img = $trad[0]['attachment'];
-        unlink('storage/' . $this->t_content_img);
-        TrixAttachment::where('is_pending', 1)->delete();
+        if (count($trad) > 0) {
+            // dd("NULL");
+            $this->t_content_img = $trad[0]['attachment'];
+            unlink('storage/' . $this->t_content_img);
+            TrixAttachment::where('is_pending', 1)->delete();
+        }
         session()->flash('msgWar', 'Anda telah membatalkan aksi.');
         return redirect()->route('showTaniPage');
     }
