+++ app/Http/Livewire/TentangBKK.php
@@ -6,6 +6,7 @@ use Livewire\Component;
 use App\Models\IdentitasSekolah;
 use App\Models\Blog;
 use Te7aHoudini\LaravelTrix\Models\TrixAttachment;
+use App\Models\ImgCarIdenSekolah;
 
 class TentangBKK extends Component
 {
@@ -68,9 +69,12 @@ class TentangBKK extends Component
     {
         // dd(request('blog-trixFields'));
         $trad = TrixAttachment::where('is_pending', 1)->get();
-        $this->t_content_img = $trad[0]['attachment'];
-        unlink('storage/' . $this->t_content_img);
-        TrixAttachment::where('is_pending', 1)->delete();
+        if (count($trad) > 0) {
+            // dd("NULL");
+            $this->t_content_img = $trad[0]['attachment'];
+            unlink('storage/' . $this->t_content_img);
+            TrixAttachment::where('is_pending', 1)->delete();
+        }
         session()->flash('msgWar', 'Anda telah membatalkan aksi.');
         return redirect()->route('showTbkkPage');
     }
@@ -80,12 +84,18 @@ class TentangBKK extends Component
         return IdentitasSekolah::orderBy('created_at', 'DESC')->get();
     }
 
+    public function readHero($tag)
+    {
+        return ImgCarIdenSekolah::select("*")->where('kategori', $tag)->get();
+        // $this->resetPage();
+    }
 
     public function render()
     {
         return view('livewire.bkk.tentang-b-k-k', [
             'data'=> $this->read(),
             'blog' => $this->readBlog(),
+            'dataHero' => $this->readHero('header'),
         ])->layout('layouts.landingpage', [
             'data'=> $this->read(),
         ]);
