+++ app/Http/Livewire/TentangKesiswaan.php
@@ -5,6 +5,7 @@ use App\Models\IdentitasSekolah;
 use Livewire\Component;
 use App\Models\Blog;
 use Te7aHoudini\LaravelTrix\Models\TrixAttachment;
+use App\Models\ImgCarIdenSekolah;
 
 class TentangKesiswaan extends Component
 {
@@ -63,13 +64,22 @@ class TentangKesiswaan extends Component
         $this->togglePage = true;
     }
 
+    public function readHero($tag)
+    {
+        return ImgCarIdenSekolah::select("*")->where('kategori', $tag)->get();
+        // $this->resetPage();
+    }
+
     public function delete_pending()
     {
         // dd(request('blog-trixFields'));
         $trad = TrixAttachment::where('is_pending', 1)->get();
-        $this->t_content_img = $trad[0]['attachment'];
-        unlink('storage/' . $this->t_content_img);
-        TrixAttachment::where('is_pending', 1)->delete();
+        if (count($trad) > 0) {
+            // dd("NULL");
+            $this->t_content_img = $trad[0]['attachment'];
+            unlink('storage/' . $this->t_content_img);
+            TrixAttachment::where('is_pending', 1)->delete();
+        }
         session()->flash('msgWar', 'Anda telah membatalkan aksi.');
         return redirect()->route('showTSiswaPage');
     }
@@ -83,6 +93,7 @@ class TentangKesiswaan extends Component
         return view('livewire.kesiswaan.tentang-kesiswaan', [
             'data'=> $this->read(),
             'blog' => $this->readBlog(),
+            'dataHero' => $this->readHero('header'),
         ])->layout('layouts.landingpage', [
             'data'=> $this->read(),
         ]);
